+++ src/server/main.py
@@ -7,7 +7,8 @@ import pandas as pd
 from fastapi import FastAPI, Request, Response
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
-from fastapi.responses import HTMLResponse
+
+from server.event_model import stress, excess
 
 PLACES = [
     "Africa",
@@ -76,17 +77,49 @@ df["date"] = pd.to_datetime(
 del df["Date Start"]
 CONTAINMENT_MEAS = df
 
+LINES = pd.read_csv(
+    "https://storage.googleapis.com/static-covid/static/line-data-v2.csv?cache_bump=2"
+)
+
+STARTDATE = pd.to_datetime("03/14/2020", format="%m/%d/%Y")
+
+# TODO: Does it break anything to replace PLACES above?
+
+PLACES_GV = {
+    "Egypt": 9.7e7,
+    "China": 1.4e9,
+    "Hong Kong": 7.4e6,
+    "India": 1.3e9,
+    "Russia": 1.44e8,
+    "South Korea": 5.1e7,
+    "Singapore": 5.6e6,
+    "Japan": 1.28e8,
+    "Indonesia": 2.64e8,
+    "Dubai": 3.3e6,
+    "Iran": 8.1e7,
+    "US": 3.3e8,
+    "Australia": 2.5e7,
+    "France": 6.7e7,
+    "Italy": 6e7,
+    "Germany": 8.3e7,
+    "Spain": 4.7e7,
+    "Netherlands": 1.7e7,
+    "Switzerland": 8.5e6,
+    "United Kingdom": 6.6e7,
+    "Czech Republic": 1.1e7,
+}
+
+LINES["Country"].unique()
+
 
 class Place:
     def __init__(self, name):
         self.name = name
 
 
-@app.get("/")
-async def index(request: Request) -> Response:
-    """TODO: this is the main hompage, should have a bubble map which should
-    link ot the /model"""
-    return templates.TemplateResponse("index.html", {"request": request},)
+@app.get("/case-map")
+async def case_map(request: Request) -> Response:
+    return templates.TemplateResponse("case-map.html", {"request": request},)
 
 
 @app.get("/request-calculation")
@@ -102,7 +135,7 @@ async def request_calculation(request: Request) -> Response:
     )
 
 
-@app.get("/model")
+@app.get("/")
 async def model(request: Request, country: str = "USA") -> Response:
     """serve the main model visualization"""
     arguments = {"country": country} if country else {}
@@ -114,7 +147,7 @@ async def model(request: Request, country: str = "USA") -> Response:
 @app.get("/request-event-evaluation")
 async def request_event_evaluation(request: Request) -> Response:
 
-    places = [Place(place) for place in PLACES]
+    places = [Place(place) for place in PLACES_GV.keys()]
 
     return templates.TemplateResponse(
         "request-event-evaluation.html",
@@ -125,20 +158,34 @@ async def request_event_evaluation(request: Request) -> Response:
 @app.get("/result-event-evaluation")
 async def result_event_evaluation(
     request: Request,
-    place: str = "USA",
-    number: int = 10,
-    datepicker: str = "02/02/2017",
+    place: str,
+    datepicker: str,
+    length: str,
+    contactSize: str,
+    controlStrength: str,
+    peopleCount: str,
 ) -> Response:
-    # TODO - implement the calculations based on parameters from the request-event-evaluation
-    # TODO: Use real data model here, also use the real values from the forms, parameters are ignored
+
+    strength = {"none": 0, "weak": 0.3, "moderate": 0.4, "strong": 0.5}[controlStrength]
+    num = {"1-10": 5, "10-100": 50, "100-1000": 500, "1000+": 5000}[peopleCount]
+
+    # Wild stab at transmission probabilities
+    transmission = {"hours": 0.02, "day": 0.05, "days": 0.08}[length]
+    effectiveContacts = {"little": 3 / 2, "inbetween": 5 / 3, "lot": 2}[contactSize]
+
     class Place:
         def __init__(self, name):
             self.name = name
-            self.population = 4000000
-            self.gleamviz_predictions = pd.Series(  # From gleamviz
-                np.logspace(2, 6, num=31),
-                index=pd.date_range(start=date.today(), periods=31),
-            )
+            self.population = PLACES_GV[name]
+            cols = [i for i in LINES.columns if i.startswith("Cumulative")]
+            timesteps = (
+                LINES[
+                    (LINES["Country"] == name) & (LINES["Mitigation"] == strength)
+                ].set_index("Timestep")[cols]
+            ) / 1000
+            self.gleamviz_predictions = timesteps.mean(axis=1)
+            self.gleamviz_upper = timesteps.max(axis=1)
+            self.gleamviz_lower = timesteps.min(axis=1)
 
     class Data:
         def __getitem__(self, name):
@@ -146,22 +193,85 @@ async def result_event_evaluation(
 
     data = Data()
 
-    # The following is not a mock, but until data are fixed, it is irrelevant
+    event_index = (pd.to_datetime(datepicker, format="%m/%d/%Y") - STARTDATE).days
+
+    dispControlStrength = controlStrength.replace("none", "no")
+
     place_data = data[place]
+
     try:
-        median = place_data.gleamviz_predictions.loc[datepicker]
-        fraction = median / place_data.population
-        probability = (1 - (1 - fraction) ** number) * 100  # in %
+        medianFraction = place_data.gleamviz_predictions.loc[event_index]
+        minFraction = place_data.gleamviz_lower.loc[event_index]
+        maxFraction = place_data.gleamviz_upper.loc[event_index]
+        minCases, maxCases, medianCases = tuple(
+            int(place_data.population * i)
+            for i in (minFraction, maxFraction, medianFraction)
+        )
+        minProbability = (1 - (1 - minFraction) ** num) * 100  # in %
+        maxProbability = (1 - (1 - maxFraction) ** num) * 100
+        # Not currently displayed
+        # minExpected = (
+        #     transmission * minFraction * num ** effectiveContacts
+        # )
+        # maxExpected = (
+        #   transmission * maxFraction * num ** effectiveContacts
+        # )
     except KeyError:
-        probability = "unknown"
+        minProbability = -99999
+        maxProbability = -99999
+        medianCases = -99999
+        fraction = -99999
+        minCases = -99999
+        maxCases = -99999
+        minExpected = -99999
+        maxExpected = -99999
+        minFraction = 0.5  # TODO: MOCKED! It was undefined in some cases
+        maxFraction = 0.5  # TODO: MOCKED! It was undefined in some cases
+
+    if strength >= 0.5:
+        minStress = "Not applicable"
+        maxStress = "Not applicable"
+    else:
+        stresses = [
+            stress(fraction * a, strength + b)
+            for a in [10, 1, 0.1]
+            for b in [0, 0.1, -0.1]
+            for fraction in [minFraction, maxFraction]
+        ]
+        minStress = min(stresses)
+        maxStress = max(stresses)
+
+    excesses = [
+        excess(fraction * a, strength + b)
+        for a in [10, 1, 0.1]
+        for b in [0, 0.1, -0.1]
+        for fraction in [minFraction, maxFraction]
+    ]
+    minExcess = min(excesses)
+    maxExcess = max(excesses)
+
     return templates.TemplateResponse(
         "result-event-evaluation.html",
         {
             "request": request,
+            "controlStrength": dispControlStrength,
             "datepicker": datepicker,
-            "number": number,
+            "peopleCount": num,
             "place": place,
-            "probability": probability,
+            "minProbability": minProbability,
+            "maxProbability": maxProbability,
+            # "minExpected": minExpected,
+            # "maxExpected": maxExpected,
+            "minExcess": minExcess,
+            "maxExcess": maxExcess,
+            "minStress": minStress,
+            "maxStress": maxStress,
+            "medianCases": medianCases,
+            "minCases": minCases,
+            "maxCases": maxCases,
+            "minPrevalence": minFraction * 100,
+            "eventDate": datepicker,
+            "startDate": STARTDATE,
         },
     )