+++ CHANGELOG.md
@@ -47,6 +47,8 @@ Here is a template for new release sections
 ### Changed
 - Template for `model` (label "Request model") (#62, #52)
 - Version of jQuery (also add it as static file) (#88)
+- Model for calculating event risk (#11)
+- Template for `request-event-evaluation` (#11)
 - Description text below the graph on `model` view (#138)
 
 ### Removed